Packs Notification-Related content into src/notifications for readability
